backend/requirements.txt
fastapi
uvicorn
firebase-admin
google-genai
python-dotenv
pydantic
